Add initial Repertoire client with builder
Add a Repertoire client package, rubin.repertoire, that is managed
as part of the same workspace. Define the initial client-facing
model and configuration model, and a builder class that generates
the service discovery results from the configuration. Add a simple
test of the builder and run it from the parent package test suite.