Add support for discovery of InfluxDB databases
Add a separate Repertoire service route to return connection
information for an InfluxDB database. This route will be authenticated
since it includes the username and password in the response.
Refactor the FastAPI dependencies to allow dynamic discovery of the
root URL of the API and configuring the root path to secrets, and
regenerate discovery information if the underlying configuration has
changed.

Include a mapping of InfluxDB database names to the URLs for the
connection information in the main service discovery information.

Add builder support for generating InfluxDB database connection
information both with and without credentials. The version without
credentials will be used later for static service discovery. Add
client support for getting a list of InfluxDB database and retrieving
their authentication information with credentials.