Convert service URLs to rich objects
Rather than returning mappings of service name to (optional) dataset
name to a simple base URL, change the final value to a full object
with a `url` key holding the URL. This creates room for future
expansion of the data that's returned.

Add an `openapi` key holding the OpenAPI schema, if available, to data
and internal services, and add the corresponding rule configuration
to template that URL.

Rename the `urls` field to `services` now that it potentially contains
more information than just URLs.