Add support for HiPS and HiPS lists
Add explicit configuration for HiPS. Construct the URLs to the
HiPS lists and return them as part of the dataset records, rather
than treating HiPS as a discoverable service. This follows the
theme of Butler of treating URLs to specific configuration or
discovery services for a dataset as separate from regular service
discovery.

Move support for generating HiPS list files from properties files
of multiple bands of HiPS surveys from datalinker into Repertoire.
Keep support for the legacy HiPS list path (without a dataset name
in the URL), but make it optional so that it can be easily removed
in the future for testing.