Allow more disabled HiPS configuration
Ignore the HiPS configuration if the relevant dataset is not in the
list of available datasets, for both the legacy and current HiPS list
URLs. Allow the legacy HiPS configuration to be partially present but
with no dataset set. This configuration flexibility allows
simplification of the Phalanx configuration by moving most of the HiPS
configuration into `values.yaml` and controlling its use through
adjusting `availableDatasets` in `values-<environment>.yaml`.