Add support for API versions
For data and internal services, allow explicit configuration of a
mapping of API versions, and include the result in the service
discovery output. Each API version currently must have a URL and may
have an IVOA standardID.

Add new client APIs `versions_for_data` and `versions_for_internal`
to retrieve the list of versions for a given API.

Add new optional `version` parameter to `url_for_data` and
`url_for_internal` to request the URL for a specific API version.