Restructure data service information
On further reflection, rather than grouping data services with
other services but using a different structure to add the dataset
identifiers, put all the service entries for a dataset under the
entry for that dataset. This avoids listing datasets in two places
in the model and structurally ensures there cannot be data services
for nonexistent datasets.

Treat HiPS like any other data service and add it to the list of
data services, instead of putting the HiPS list URL in a separate
`hips_list` key directly under the dataset.